+++ pinot-core/src/main/java/org/apache/pinot/core/data/table/IndexedTableResizer.java
@@ -18,22 +18,258 @@
  */
 package org.apache.pinot.core.data.table;
 
+import com.google.common.annotations.VisibleForTesting;
+import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.PriorityQueue;
+import java.util.Set;
+import org.apache.commons.collections.comparators.ComparableComparator;
+import org.apache.pinot.common.request.AggregationInfo;
+import org.apache.pinot.common.request.SelectionSort;
+import org.apache.pinot.common.utils.DataSchema;
+import org.apache.pinot.core.query.aggregation.function.AggregationFunction;
+import org.apache.pinot.core.query.aggregation.function.AggregationFunctionUtils;
 
 
 /**
- * Helper class for trimming and sorting records in the IndexedTable
+ * Helper class for trimming and sorting records in the IndexedTable, based on the order by information
  */
-abstract class IndexedTableResizer {
+class IndexedTableResizer {
+
+  private List<OrderByInfo> _orderByInfos;
+  private Comparator<IntermediateRecord> _intermediateRecordComparator;
+  private int _numOrderBy;
+
+  IndexedTableResizer(DataSchema dataSchema, List<AggregationInfo> aggregationInfos, List<SelectionSort> orderBy) {
+
+    int numAggregations = aggregationInfos.size();
+    int numKeyColumns = dataSchema.size() - numAggregations;
+
+    Map<String, Integer> keyIndexMap = new HashMap<>();
+    for (int i = 0; i < numKeyColumns; i++) {
+      keyIndexMap.put(dataSchema.getColumnName(i), i);
+    }
+
+    Map<String, Integer> aggregationColumnToIndex = new HashMap<>();
+    Map<String, AggregationInfo> aggregationColumnToInfo = new HashMap<>();
+    for (int i = 0; i < numAggregations; i++) {
+      AggregationInfo aggregationInfo = aggregationInfos.get(i);
+      String aggregationColumn = AggregationFunctionUtils.getAggregationColumnName(aggregationInfo);
+      aggregationColumnToIndex.put(aggregationColumn, i);
+      aggregationColumnToInfo.put(aggregationColumn, aggregationInfo);
+    }
+
+    _numOrderBy = orderBy.size();
+    _orderByInfos = new ArrayList<>(_numOrderBy);
+    Comparator[] comparators = new Comparator[_numOrderBy];
+
+    for (int i = 0; i < _numOrderBy; i++) {
+      SelectionSort selectionSort = orderBy.get(i);
+      String column = selectionSort.getColumn();
+
+      if (keyIndexMap.containsKey(column)) {
+        int index = keyIndexMap.get(column);
+        _orderByInfos.add(createOrderByInfoForKeyColumn(index));
+      } else if (aggregationColumnToIndex.containsKey(column)) {
+        int index = aggregationColumnToIndex.get(column);
+        _orderByInfos.add(createOrderByInfoForAggregationColumn(index, aggregationColumnToInfo.get(column)));
+      } else {
+        throw new IllegalStateException("Could not find column " + column + " in data schema");
+      }
+
+      comparators[i] = Comparator.naturalOrder();
+      if (!selectionSort.isIsAsc()) {
+        comparators[i] = comparators[i].reversed();
+      }
+    }
+
+    _intermediateRecordComparator = (o1, o2) -> {
+
+      for (int i = 0; i < _numOrderBy; i++) {
+        int result = comparators[i].compare(o1._values[i], o2._values[i]);
+        if (result != 0) {
+          return result;
+        }
+      }
+      return 0;
+    };
+  }
 
   /**
-   * Trim recordsMap to trimToSize
+   * Constructs an IntermediateRecord from Record
+   * The IntermediateRecord::key is the same Record::key
+   * The IntermediateRecord::values contains only the order by columns, in the query's sort sequence
+   * For aggregation values in the order by, the final result is extracted if the intermediate result is non-comparable
    */
-  abstract void resizeRecordsMap(Map<Key, Record> recordsMap, int trimToSize);
+  @VisibleForTesting
+  IntermediateRecord getIntermediateRecord(Record record) {
+    Comparable[] intermediateRecordValues = new Comparable[_numOrderBy];
+
+    Object[] keyColumns = record.getKey().getColumns();
+    Object[] aggregationColumns = record.getValues();
+    for (int i = 0; i < _numOrderBy; i++) {
+      OrderByInfo orderByInfo = _orderByInfos.get(i);
+      Comparable comparable;
+      if (orderByInfo.getOrderByType().equals(OrderByType.KEY_COLUMN)) {
+        Object keyColumn = keyColumns[orderByInfo.getIndex()];
+        comparable = (Comparable) keyColumn; // FIXME: is this the right way to get Comparable? will it work for BYTES?
+      } else {
+        Object aggregationColumn = aggregationColumns[orderByInfo.getIndex()];
+        AggregationFunction aggregationFunction = orderByInfo.getAggregationFunction();
+        if (!aggregationFunction.isIntermediateResultComparable()) {
+          comparable = aggregationFunction.extractFinalResult(aggregationColumn);
+        } else {
+          comparable = (Comparable) aggregationColumn;
+        }
+      }
+      intermediateRecordValues[i] = comparable;
+    }
+    return new IntermediateRecord(record.getKey(), intermediateRecordValues);
+  }
 
   /**
-   * Return a sorted list of records from the recordsMap
+   * Trim recordsMap to trimToSize, based on order by information
    */
-  abstract List<Record> sortRecordsMap(Map<Key, Record> recordsMap);
+  void resizeRecordsMap(Map<Key, Record> recordsMap, int trimToSize) {
+
+    int numRecordsToEvict = recordsMap.size() - trimToSize;
+
+    if (numRecordsToEvict > 0) {
+      int size;
+      Comparator<IntermediateRecord> comparator;
+
+      if (numRecordsToEvict < trimToSize) { // num records to evict is smaller than num records to retain
+        size = numRecordsToEvict;
+        // make PQ of records to evict
+        comparator = _intermediateRecordComparator;
+      } else { // num records to retain is smaller than num records to evict
+        size = trimToSize;
+        // make PQ of records to retain
+        comparator = _intermediateRecordComparator.reversed();
+      }
+
+      PriorityQueue<IntermediateRecord> priorityQueue =
+          new PriorityQueue<>(size, comparator);
+
+      for (Record record : recordsMap.values()) {
+
+        IntermediateRecord intermediateRecord = getIntermediateRecord(record);
+        if (priorityQueue.size() < size) {
+          priorityQueue.offer(intermediateRecord);
+        } else {
+          IntermediateRecord peek = priorityQueue.peek();
+          if (priorityQueue.comparator().compare(peek, intermediateRecord) < 0) {
+            priorityQueue.poll();
+            priorityQueue.offer(intermediateRecord);
+          }
+        }
+      }
+
+      if (numRecordsToEvict < trimToSize) { // PQ contains records to evict
+        for (IntermediateRecord evictRecord : priorityQueue) {
+          recordsMap.remove(evictRecord._key);
+        }
+      } else { // PQ contains records to retain
+        ObjectOpenHashSet<Key> keysToRetain = new ObjectOpenHashSet<>(priorityQueue.size());
+        for (IntermediateRecord retainRecord : priorityQueue) {
+          keysToRetain.add(retainRecord._key);
+        }
+        recordsMap.keySet().retainAll(keysToRetain);
+      }
+    }
+  }
+
+  /**
+   * Sort the records in the recordsMap according to order by
+   */
+  List<Record> sortRecordsMap(Map<Key, Record> recordsMap) {
+    int numRecords = recordsMap.size();
+    if (numRecords == 0) {
+      return Collections.emptyList();
+    }
+    List<Record> sortedRecords = new ArrayList<>(numRecords);
+    List<IntermediateRecord> intermediateRecords = new ArrayList<>(numRecords);
+    for (Record record : recordsMap.values()) {
+      intermediateRecords.add(getIntermediateRecord(record));
+    }
+    intermediateRecords.sort(_intermediateRecordComparator);
+    for (IntermediateRecord intermediateRecord : intermediateRecords) {
+      sortedRecords.add(recordsMap.get(intermediateRecord._key));
+    }
+    return sortedRecords;
+  }
+
+
+  /**
+   * Helper class to store a subset of Record fields
+   * IntermediateRecord is derived from a Record
+   * Some of the main properties of an IntermediateRecord are:
+   *
+   * 1. Key in IntermediateRecord is expected to be identical to the one in the Record
+   * 2. For values, IntermediateRecord should only have the columns needed for order by
+   * 3. Inside the values, the columns should be ordered by the order by sequence
+   * 4. For order by on aggregations, final results should extracted if the intermediate result is non-comparable
+   */
+  @VisibleForTesting
+  static class IntermediateRecord {
+    final Key _key;
+    final Comparable[] _values;
+
+    IntermediateRecord(Key key, Comparable[] values) {
+      _key = key;
+      _values = values;
+    }
+  }
+
+  /**
+   * Enum to differentiate between different types of order by
+   * Order by can be on the key column (group by values or distinct), or on the aggregation column
+   */
+  private enum OrderByType {
+    KEY_COLUMN, // order by on a key column (e.g. group by or distinct)
+    AGGREGATION_COLUMN // order by on aggregation (e.g. aggregations in aggregation only or group by)
+  }
+
+  /**
+   * Defines the information needed for one SelectionSort of an order by clause
+   */
+  private static class OrderByInfo {
+    final OrderByType _orderByType;
+    final int _index; // if orderByType = KEY_COLUMN, index in the Record::key, else index in the Record::value
+    final AggregationFunction _aggregationFunction; // required if orderByType = AGGREGATION_COLUMN
+
+    OrderByInfo(OrderByType orderByType, int index, AggregationFunction aggregationFunction) {
+      _orderByType = orderByType;
+      _index = index;
+      _aggregationFunction = aggregationFunction;
+    }
+
+    OrderByType getOrderByType() {
+      return _orderByType;
+    }
+
+    int getIndex() {
+      return _index;
+    }
+
+    AggregationFunction getAggregationFunction() {
+      return _aggregationFunction;
+    }
+  }
+
+  private OrderByInfo createOrderByInfoForKeyColumn(int index) {
+    return new OrderByInfo(OrderByType.KEY_COLUMN, index, null);
+  }
+
+  private OrderByInfo createOrderByInfoForAggregationColumn(int index, AggregationInfo aggregationInfo) {
+    AggregationFunction aggregationFunction =
+        AggregationFunctionUtils.getAggregationFunctionContext(aggregationInfo).getAggregationFunction();
+    return new OrderByInfo(OrderByType.AGGREGATION_COLUMN, index, aggregationFunction);
+  }
 }
